change exported comp name
