+++ apps/nextjs/content/pages/toys.mdx
@@ -2,16 +2,14 @@
 title: Toys
 ---
 
-{" "}
-
-<BoxDrawn.Box
+<BoxDrawn
   rows={5}
   variant="classic"
   shadow
   className="text-2xl md:text-3xl not-prose font-bold text-center max-w-[28ch] mx-auto mb-14"
 >
   <h1>Toys I've made...</h1>
-</BoxDrawn.Box>
+</BoxDrawn>
 
 | Name                                                                                                                                       | Description                                                                                                                                         | Repo                                                                                       |
 | ------------------------------------------------------------------------------------------------------------------------------------------ | --------------------------------------------------------------------------------------------------------------------------------------------------- | ------------------------------------------------------------------------------------------ |
